test(netcdf): move the netcdf_samples suite under tests/netcdf as samples/
Relocate the standalone tests/netcdf_samples corpus suite (34 files, its
own conftest with the SAMPLES registry + pytest_generate_tests) into
tests/netcdf/samples/ so all NetCDF tests live under one tree. Repoint the
13 intra-suite 'from tests.netcdf_samples.conftest import ...' imports to
tests.netcdf.samples.conftest.

The suite keeps its own conftest, so its parametrization and data-gated
skips are unchanged; tests/netcdf/conftest.py has no autouse fixtures or
hooks and no fixture-name overlap, so inheriting it introduces no new
behaviour. Pure relocation: 725 sample tests collect unchanged (2315 total
under tests/netcdf/), no bodies edited.
