+++ tests/netcdf/samples/test_array_io.py
