+++ tests/netcdf/samples/test_dimensions_coords.py
