+++ tests/netcdf/samples/test_global_attributes.py
