+++ tests/netcdf/samples/test_groups.py
