+++ tests/netcdf/samples/test_read_and_open.py
