+++ tests/netcdf/samples/test_reduce.py
@@ -3,7 +3,7 @@
 import pytest
 
 from pyramids.netcdf import NetCDF
-from tests.netcdf_samples.conftest import RHUM
+from tests.netcdf.samples.conftest import RHUM
 
 pytestmark = pytest.mark.core
