+++ tests/netcdf/samples/test_roundtrip_bytes.py
