+++ tests/netcdf/samples/test_smoke_all_files.py
