+++ tests/netcdf/samples/test_ugrid.py
@@ -3,7 +3,7 @@
 import pytest
 
 from pyramids.netcdf import NetCDF, UgridDataset
-from tests.netcdf_samples.conftest import MESH
+from tests.netcdf.samples.conftest import MESH
 
 pytestmark = pytest.mark.core
