+++ tests/netcdf/samples/test_variables_access.py
